修复写模式下可能会产生的 bug
