--- packages/mip-sandbox/test/spec/global-detect.spec.js
@@ -1,26 +1,0 @@
-/**
- * @file global-detect.spec.js
- * @author clark-t (clarktanglei@163.com)
- */
-
-/* globals describe, it */
-// var chai = require('chai')
-// var expect = chai.expect
-// var mark = require('../../lib/global-mark')
-// // var detect = require('../../lib/global-detect')
-// var acorn = require('acorn-dynamic-import').default
-
-// describe.only('import()', function () {
-//   it('should be pass', function () {
-//     var code = `
-//       import('path/to/remote')
-//     `
-
-//     var ast = acorn.parse(code, {plugins: {dynamicImport: true}})
-
-//     // console.log(mark(code))
-//     // console.log(JSON.stringify(ast, null, 2))
-//     // expect(acorn.parse(code, {plugins: {dynamicImport: true}})).to.not.throw()
-//     // expect(detect(code)).to.not.throw()
-//   })
-// })
