+++ packages/mip-sandbox/test/spec/sandbox.spec.js
@@ -46,6 +46,10 @@ describe('sandbox', function () {
     expect(sandbox.this(window)).to.not.be.equal(window)
     expect(sandbox.this(window)).to.be.equal(sandbox)
     expect(sandbox.this(document)).to.be.equal(sandbox.document)
+    expect(sandbox.document.cookie).to.be.equal(document.cookie)
+    expect(Object.keys(sandbox.strict.document)).to.be.deep.equal(['cookie', 'domain'])
+    expect(sandbox.strict.document.cookie).to.be.equal(document.cookie)
+    expect(sandbox.strict.MIP.viewer.isIframed).to.be.equal(MIP.viewer.isIframed)
     expect(sandbox.strict.this(window)).to.be.equal(sandbox.strict)
     expect(sandbox.strict.this(document)).to.be.equal(sandbox.strict.document)
   })
